Added active badge

+++ app/src/main/java/com/ken/myapplication/data/AcStats.kt
@@ -2,7 +2,7 @@ package com.ken.myapplication.data
 
 data class AcStats(
     val acQuestionCount: Int,
-    val acRate: Int,
+    val acRate: Double,
     val acSubmissionCount: Int,
     val totalSubmissionCount: Int
 )
